Refactor parsers to prepare for v2
